pyproject.toml
[project]
name = "pdf-extractor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.27.0",
    "jupyterlab>=4.3.4",
    "langchain-community>=0.3.14",
    "langchain-openai>=0.3.1",
    "langgraph>=0.2.64",
    "mypy>=1.14.1",
    "openai>=1.59.9",
    "pydantic>=2.10.5",
    "pypdf>=5.1.0",
    "streamlit>=1.41.1",
]
